+++ modules/gapi/src/backends/cpu/gcpucore.cpp
@@ -462,30 +462,6 @@ GAPI_OCV_KERNEL(GCPUMerge4, cv::gapi::core::GMerge4)
     }
 };
 
-GAPI_OCV_KERNEL(GCPUResize, cv::gapi::core::GResize)
-{
-    static void run(const cv::Mat& in, cv::Size sz, double fx, double fy, int interp, cv::Mat &out)
-    {
-        cv::resize(in, out, sz, fx, fy, interp);
-    }
-};
-
-GAPI_OCV_KERNEL(GCPUResizeP, cv::gapi::core::GResizeP)
-{
-    static void run(const cv::Mat& in, cv::Size out_sz, int interp, cv::Mat& out)
-    {
-        int inH = in.rows / 3;
-        int inW = in.cols;
-        int outH = out.rows / 3;
-        int outW = out.cols;
-        for (int i = 0; i < 3; i++) {
-            auto in_plane = in(cv::Rect(0, i*inH, inW, inH));
-            auto out_plane = out(cv::Rect(0, i*outH, outW, outH));
-            cv::resize(in_plane, out_plane, out_sz, 0, 0, interp);
-        }
-    }
-};
-
 GAPI_OCV_KERNEL(GCPURemap, cv::gapi::core::GRemap)
 {
     static void run(const cv::Mat& in, const cv::Mat& x, const cv::Mat& y, int a, int b, cv::Scalar s, cv::Mat& out)
@@ -775,8 +751,6 @@ cv::GKernelPackage cv::gapi::core::cpu::kernels()
          , GCPUInRange
          , GCPUSplit3
          , GCPUSplit4
-         , GCPUResize
-         , GCPUResizeP
          , GCPUMerge3
          , GCPUMerge4
          , GCPURemap
