+++ modules/gapi/test/cpu/gapi_core_tests_cpu.cpp
@@ -301,42 +301,6 @@ INSTANTIATE_TEST_CASE_P(Split4TestCPU, Split4Test,
                                 Values(CV_8UC1),
                                 Values(CORE_CPU)));
 
-INSTANTIATE_TEST_CASE_P(ResizeTestCPU, ResizeTest,
-                        Combine(Values( CV_8UC1, CV_8UC3, CV_16UC1, CV_16SC1, CV_32FC1 ),
-                                Values(cv::Size(1280, 720),
-                                       cv::Size(640, 480),
-                                       cv::Size(128, 128)),
-                                Values(-1),
-                                Values(CORE_CPU),
-                                Values(AbsSimilarPoints(2, 0.05).to_compare_obj()),
-                                Values(cv::INTER_NEAREST, cv::INTER_LINEAR, cv::INTER_AREA),
-                                Values(cv::Size(64,64),
-                                       cv::Size(30,30))));
-
-INSTANTIATE_TEST_CASE_P(ResizePTestCPU, ResizePTest,
-                        Combine(Values( CV_8UC1, CV_8UC3, CV_16UC1, CV_16SC1, CV_32FC1 ),
-                                Values(cv::Size(1280, 720),
-                                       cv::Size(640, 480),
-                                       cv::Size(128, 128)),
-                                Values(-1),
-                                Values(CORE_CPU),
-                                Values(AbsSimilarPoints(2, 0.05).to_compare_obj()),
-                                Values(cv::INTER_LINEAR),
-                                Values(cv::Size(64,64),
-                                       cv::Size(30,30))));
-
-INSTANTIATE_TEST_CASE_P(ResizeTestCPU, ResizeTestFxFy,
-                        Combine(Values( CV_8UC1, CV_8UC3, CV_16UC1, CV_16SC1, CV_32FC1 ),
-                                Values(cv::Size(1280, 720),
-                                       cv::Size(640, 480),
-                                       cv::Size(128, 128)),
-                                Values(-1),
-                                Values(CORE_CPU),
-                                Values(AbsSimilarPoints(2, 0.05).to_compare_obj()),
-                                Values(cv::INTER_NEAREST, cv::INTER_LINEAR, cv::INTER_AREA),
-                                Values(0.5, 0.1),
-                                Values(0.5, 0.1)));
-
 INSTANTIATE_TEST_CASE_P(Merge3TestCPU, Merge3Test,
                         Combine(Values(CV_8UC1),
                                 Values(cv::Size(1280, 720),
